Move *Ref and IntoEdge traits to the base submodule
The plan is to merge vertices/edges traits and create a new hierarchy.
Since the vertex/edge dichotomy will no longer exist in the trait
hierarchy, it makes sense to move these basic traits to one place.
