+++ gryf/src/core/edges.rs
@@ -5,18 +5,11 @@ use thiserror::Error;
 use crate::common::CompactIdMap;
 
 use super::{
-    base::GraphBase,
-    id::{GraphIdTypes, IdType, IntegerIdType},
+    base::{EdgeRef, GraphBase},
+    id::IntegerIdType,
     marker::EdgeType,
     weak::WeakRef,
 };
-
-pub trait EdgeRef<VId: IdType, EId: IdType, E> {
-    fn id(&self) -> &EId;
-    fn data(&self) -> &E;
-    fn src(&self) -> &VId;
-    fn dst(&self) -> &VId;
-}
 
 pub trait EdgesBase<Ty: EdgeType>: GraphBase {
     type EdgeIdsIter<'a>: Iterator<Item = Self::EdgeId>
@@ -178,58 +171,6 @@ pub trait EdgesWeak<E, Ty: EdgeType>: EdgesBaseWeak<Ty> {
 }
 
 pub trait MultiEdges<Ty: EdgeType>: EdgesBase<Ty> {}
-
-pub trait IntoEdge<Id: GraphIdTypes, E, Ty: EdgeType> {
-    fn unpack(self) -> (Id::VertexId, Id::VertexId, E);
-}
-
-impl<'a, VId: IdType, EId: IdType, E> EdgeRef<VId, EId, E> for (EId, &'a E, VId, VId) {
-    fn id(&self) -> &EId {
-        &self.0
-    }
-
-    fn data(&self) -> &E {
-        self.1
-    }
-
-    fn src(&self) -> &VId {
-        &self.2
-    }
-
-    fn dst(&self) -> &VId {
-        &self.3
-    }
-}
-
-impl<Id: GraphIdTypes, E, Ty: EdgeType, I: Into<Id::VertexId>> IntoEdge<Id, E, Ty> for (I, I, E) {
-    fn unpack(self) -> (Id::VertexId, Id::VertexId, E) {
-        (self.0.into(), self.1.into(), self.2)
-    }
-}
-
-impl<Id: GraphIdTypes, E: Clone, Ty: EdgeType, I: Into<Id::VertexId> + Clone> IntoEdge<Id, E, Ty>
-    for &(I, I, E)
-{
-    fn unpack(self) -> (Id::VertexId, Id::VertexId, E) {
-        (self.0.clone().into(), self.1.clone().into(), self.2.clone())
-    }
-}
-
-impl<Id: GraphIdTypes, E: Default, Ty: EdgeType, I: Into<Id::VertexId>> IntoEdge<Id, E, Ty>
-    for (I, I)
-{
-    fn unpack(self) -> (Id::VertexId, Id::VertexId, E) {
-        (self.0.into(), self.1.into(), E::default())
-    }
-}
-
-impl<Id: GraphIdTypes, E: Default, Ty: EdgeType, I: Into<Id::VertexId> + Clone> IntoEdge<Id, E, Ty>
-    for &(I, I)
-{
-    fn unpack(self) -> (Id::VertexId, Id::VertexId, E) {
-        (self.0.clone().into(), self.1.clone().into(), E::default())
-    }
-}
 
 macro_rules! deref_edges_base {
     ($($ref_kind:tt)*) => {
